--- src/main/java/umc/cockple/demo/domain/notification/enums/outbox/NotificationPushChannel.java
@@ -1,5 +1,0 @@
-package umc.cockple.demo.domain.notification.enums.outbox;
-
-public enum NotificationPushChannel {
-    FCM
-}
