--- src/main/java/umc/cockple/demo/domain/notification/service/outbox/ClaimedNotificationPushOutbox.java
@@ -1,9 +1,0 @@
-package umc.cockple.demo.domain.notification.service.outbox;
-
-import umc.cockple.demo.domain.notification.domain.outbox.NotificationPushOutbox;
-
-public record ClaimedNotificationPushOutbox(
-        NotificationPushOutbox outbox,
-        String claimToken
-) {
-}
