+++ src/main/java/umc/cockple/demo/domain/push/enums/NotificationPushChannel.java
@@ -1,0 +1,5 @@
+package umc.cockple.demo.domain.push.enums;
+
+public enum NotificationPushChannel {
+    FCM
+}
