+++ src/main/java/umc/cockple/demo/domain/push/service/ClaimedNotificationPushOutbox.java
@@ -1,0 +1,9 @@
+package umc.cockple.demo.domain.push.service;
+
+import umc.cockple.demo.domain.push.domain.NotificationPushOutbox;
+
+public record ClaimedNotificationPushOutbox(
+        NotificationPushOutbox outbox,
+        String claimToken
+) {
+}
