+++ src/test/java/umc/cockple/demo/domain/push/listener/ChatNotificationEventListenerTest.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.notification.listener;
+package umc.cockple.demo.domain.push.listener;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -11,13 +11,13 @@ import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 import umc.cockple.demo.domain.chat.enums.ChatRoomType;
 import umc.cockple.demo.domain.chat.events.ChatNotificationEvent;
-import umc.cockple.demo.domain.notification.service.outbox.NotificationPushOutboxService;
+import umc.cockple.demo.domain.push.service.NotificationPushOutboxService;
 
 import java.lang.reflect.Method;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.then;
 import static org.mockito.BDDMockito.willThrow;
 import static org.mockito.Mockito.only;
@@ -48,29 +48,28 @@ class ChatNotificationEventListenerTest {
     }
 
     @Test
-    @DisplayName("Push Outbox 저장 예외는 채팅 메시지로 전파하지 않는다")
-    void handleChatNotification_swallowsException() {
+    @DisplayName("Push Outbox 저장 예외는 채팅 메시지 트랜잭션으로 전파되어 함께 롤백된다")
+    void handleChatNotification_propagatesException() {
         // given
         ChatNotificationEvent event = event();
         willThrow(new RuntimeException("아웃박스 장애"))
                 .given(notificationPushOutboxService).enqueueChat(event);
 
-        assertThatCode(() -> listener.handleChatNotification(event))
-                .doesNotThrowAnyException();
+        assertThatThrownBy(() -> listener.handleChatNotification(event))
+                .isInstanceOf(RuntimeException.class);
     }
 
     @Test
-    @DisplayName("핸들러는 AFTER_COMMIT 트랜잭션 이벤트와 Push executor로 동작한다")
-    void handler_usesPushExecutorAndAfterCommit() throws NoSuchMethodException {
+    @DisplayName("핸들러는 BEFORE_COMMIT 트랜잭션 이벤트로 동기 동작하며 비동기가 아니다")
+    void handler_usesBeforeCommitAndNotAsync() throws NoSuchMethodException {
         Method handler = ChatNotificationEventListener.class
                 .getMethod("handleChatNotification", ChatNotificationEvent.class);
 
         Async async = handler.getAnnotation(Async.class);
-        assertThat(async).isNotNull();
-        assertThat(async.value()).isEqualTo("notificationPushExecutor");
+        assertThat(async).isNull();
 
         TransactionalEventListener transactional = handler.getAnnotation(TransactionalEventListener.class);
         assertThat(transactional).isNotNull();
-        assertThat(transactional.phase()).isEqualTo(TransactionPhase.AFTER_COMMIT);
+        assertThat(transactional.phase()).isEqualTo(TransactionPhase.BEFORE_COMMIT);
     }
 }
